refactor: LectureException exception 패키지로 이동
