+++ src/main/kotlin/com/example/hhpluslectureapply/exception/LectureException.kt
@@ -1,4 +1,4 @@
-package com.example.hhpluslectureapply.domain.lecture
+package com.example.hhpluslectureapply.exception
 
 class LectureException(override val message: String) : RuntimeException() {
